requirements.txt
flask
gunicorn
sentence-transformers
python-dotenv
flask-cors
pinecone
psutil